+++ BPC.Admin/Private/ConvertTo-DotNetCookie.ps1
@@ -1,0 +1,8 @@
+Function ConvertTo-DotNetCookie {
+    [CmdletBinding()]
+    [OutputType([System.Net.Cookie])]
+    param (
+        [string]
+        $CookieString
+    )
+}