+++ BPC.Admin/Private/GetHelloWorld.ps1
@@ -1,0 +1,3 @@
+function GetBPERPExample {
+    'Hello world'
+}